ml/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
sqlalchemy==2.0.23
python-dotenv==1.0.0
httpx==0.25.1

# ML and data processing
numpy==1.26.1
pandas==2.1.2
scikit-learn==1.3.2
torch==2.1.1
transformers==4.35.0
sentence-transformers==2.2.2

# Wikipedia and scraping
beautifulsoup4==4.12.2
wikipedia-api==0.6.0

# GitHub and API integrations
PyGithub==2.1.1
requests==2.31.0

# Testing and development
pytest==7.4.3
pytest-cov==4.1.0
black==23.10.1
flake8==6.1.0
mypy==1.6.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.6

# Utilities
tqdm==4.66.1
python-multipart==0.0.6

# AI and ML
crewai>=0.1.0
langchain>=0.0.200
google-generativeai>=0.3.0
spacy>=3.5.0
nltk>=3.8.1

# Graph database
neo4j>=5.0.0
networkx>=2.8.0

# Document processing
python-docx>=0.8.11
PyPDF2>=3.0.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.5.0

# Testing
pytest-asyncio>=0.15.0

# Development
isort>=5.9.0